# openapi.yaml
openapi: 3.0.3
info:
  title: DSA Tracker API
  version: 1.0.0
  description: |
    DSA Mastery — Editorial Paper tracker. 56 chapters, 843 unique problems,
    progress sync, LC proxy, PDF viewer. Auth via Bearer JWT.
servers:
  - url: http://localhost:3000/api
tags:
  - name: Auth
  - name: Progress
  - name: System
paths:
  /register:
    post:
      tags: [Auth]
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8, description: Min 8, letter+number }
                name: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/AuthResponse" } } } }
        "400": { description: Validation/email exists }
  /login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [email, password], properties: { email: { type: string, format: email }, password: { type: string } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/AuthResponse" } } } }
        "401": { description: Bad credentials }
        "429": { description: Locked }
  /me:
    get:
      tags: [Auth]
      summary: Current user
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { user: { $ref: "#/components/schemas/User" } } } } } }
  /progress:
    get:
      tags: [Progress]
      summary: Get progress
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: OK } }
    post:
      tags: [Progress]
      summary: Merge progress (idempotent shallow merge)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                topics: { type: object }
                questions: { type: object }
                notes: { type: object }
                topicNotes: { type: object }
                chapterNotes: { type: object }
                _time: { type: object }
                lcUser: { type: string }
                companyPrefs: { type: object, properties: { target: { type: string }, timeframe: { type: string } } }
      responses: { "200": { description: Merged } }
  /companies:
    get:
      tags: [Progress]
      summary: List companies with counts
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string }
          description: Search filter
      responses: { "200": { description: OK } }
  /company/{slug}:
    get:
      tags: [Progress]
      summary: Company questions with filters
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
        - { name: timeframe, in: query, required: false, schema: { type: string, enum: [30d, 3m, 6m, gt6m, all] } }
        - { name: difficulty, in: query, required: false, schema: { type: string, enum: [Easy, Medium, Hard] } }
        - { name: q, in: query, required: false, schema: { type: string } }
        - { name: minFreq, in: query, required: false, schema: { type: number } }
        - { name: sort, in: query, required: false, schema: { type: string, enum: [frequency, title, difficulty] } }
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: limit, in: query, required: false, schema: { type: integer } }
      responses: { "200": { description: OK }, "404": { description: Company not found } }
  /question/{lcId}/companies:
    get:
      tags: [Progress]
      summary: Companies asking a question
      parameters: [{ name: lcId, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
  /lc/{username}:
    get:
      tags: [Progress]
      summary: LeetCode proxy with 5m cache
      parameters: [{ name: username, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { source: { type: string }, data: { type: object } } } } } }
        "502": { description: Upstream failed }
  /health:
    get:
      tags: [System]
      summary: Liveness
      security: []
      responses: { "200": { description: OK } }
  /readyz:
    get:
      tags: [System]
      summary: Readiness
      security: []
      responses: { "200": { description: Ready }, "503": { description: Not ready } }
  /metrics:
    get:
      tags: [System]
      summary: Metrics
      security: []
      responses: { "200": { description: text/plain } }
  /docs:
    get:
      tags: [System]
      summary: Browsable docs
      security: []
      responses: { "200": { description: HTML } }
  /docs/openapi.yaml:
    get:
      tags: [System]
      summary: Raw spec
      security: []
      responses: { "200": { description: YAML } }

components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    User: { type: object, properties: { id: { type: string }, email: { type: string }, name: { type: string } } }
    AuthResponse: { type: object, properties: { token: { type: string }, user: { $ref: "#/components/schemas/User" } } }